- added check_file_exists() method to FileUtil; used to guard against   overwriting expensive output - mindist: circumvent sqlite <-> numpy.int64 incompatibility by   directly loading the db from the GdistData stream
git-svn-id: svn+ssh://gonzo.med.jhmi.edu/scratch/svn/woolf_repository/users/oliver/Library/GromacsWrapper@3524 df5ba8eb-4b0b-0410-8c14-c10f23b0129c

+++ gromacs/analysis/plugins.py
@@ -80,6 +80,7 @@ class Worker(FileUtils):
         _filename = self.filename(filename, ext=ext, use_my_ext=True)
         pylab.savefig(_filename)
         print "Saved figure as %(_filename)r." % vars()
+            
 
 class _CysAccessibility(Worker):
     """Analysis of Cysteine accessibility."""
@@ -163,6 +164,8 @@ class _CysAccessibility(Worker):
             groupname = 'Cys%(resid)d' % vars()
             commands = [groupname, 'OW']
             filename = self.parameters.filenames[resid]
+            if self.check_file_exists(filename, resolve='warning'):
+                continue
             print "run_g_dist: %(groupname)s --> %(filename)r" % vars()
             sys.stdout.flush()
             datafile = open(filename, 'w')
